+++ tests/cli_tests.py
@@ -5,7 +5,7 @@ from click.testing import CliRunner
 from wdc.controller.export_import import ExportType
 from wdc.controller.work_day import WdcTaskInfo
 from wdc.classes import WdcTask
-from wdc.runner import cli, task_to_printout
+from wdc.runner import cli, task_to_printout, print_info
 from freezegun import freeze_time
 
 
@@ -349,3 +349,15 @@ class ExportCommandFixture(unittest.TestCase):
         self.assertEqual('export_today.csv', call_args['file_path'])
         # Assert that JSON is selected as the export type
         self.assertEqual(ExportType.CSV, call_args['export_to'])
+
+
+class PrintHelperFixture(unittest.TestCase):
+    """
+    Unit tests for all the print testing functions in the application
+    """
+
+    @patch('builtins.print')
+    def test_info(self, mock_print):
+        print_info('Test message')
+
+        mock_print.assert_called_with('\n\x1b[38;5;0m\x1b[48;5;164minfo: Test message \x1b[0m\n')