+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Add
 - ```export``` command
 
+### Fixed
+- Tasks with no histroy do not cause error on ```info``` command
+
 ## [0.3.9]
 ### Add
 - ```amend``` command